Add multiple custom feature detectors, extractors and matchers.
We have our own inhouse k-nearest neighbor matching algorithm,
but now we can use FLANN, brute force, etc. Keep in mind that
the old inhouse SURF feature detector is not working due to
the version but also the old one is not working due to a
segmentation bug in the current version. This excludes two
feature detectors, but ORB is working, so the unit tests use
it. The old SURF detector is used for people with old
OpenCV versions.