Add a list of all available feature detectors, extractors, and matchers.
Some of these don't work due to bugs, others are impeded by incompatibilities
in implementations although the API is supposed to be the same according to
the original common interfaces documentation of OpenCV. A third problem among
them is the scaling of similarity. Some matchers work with greater distances,
while others don't. A solution would be to give this choice to the developer,
therefore "similarity" needs to be redefined as "tolerance" or "distance".