Add image logging for the template matching.
This is aimed to help any debugging with regard to image matching:
why an image wasn't matched, why a different one was matched, etc.
The black and white visualization (lighter parts of the screen are
better matches) could also be used to measure match fuzziness and
difficulty, distracting factors, and even what method to use in
order to match better.