Remove subregion functionality from image finding to improve coherence of functionality.
It can be imlemented at the Region class or some image processing module. In the
ImageFinder it was supported only by autopy which is one of the least usable
finding methods. ImageFinder is now solely about finding an image and contains
multiple methods of doing this which already makes its code a bit too large.