Add a hybrid method for finding an image using feature matching in template-like manner.
The purpose of this method is to reduce the difficulty of feature matching when the
sizes of the needle and haystack images are too disproportional. It can also be used
as a basis for feature matching find_all method since it performs feature matching
for each subregion and can thus identify multiple matches. Although possible, try
not to divide the haystack into too small regions with too small translation distance
since there are internal issues with memory leaks, i.e. use moderately.