Implement a RigidMatcher that uses information on location in order to better select reliable matches.
By restricting the kinds of possible transformations, all features can be selected based on their position relative to each other.
This is a solution of the "spaghetti monster" problem of some feature matching where the final matches rely solely on descriptor similarity.
Current improvements are being made with regard to performance through heuristics to guide the search for the best set of matches.