Perform all image matching unit tests in full screen mode.
The purpose of this is to reduce variability due to undefined areas of the screen (everything outside
the window of the shown needle image or Qt4 application). All tests need to be isolated from external
conditions which in this case are other opened applications or windows visible on the screen.