Be explicit about calls in the string methods of each object.
In a string method the intention of usage is clear - to print
or convert to string. Despite this python complains when
calling an object in a string method, implicitly assuming
this intention and expecting that the interpreter will call the
string methods of the internal objects used. This results in
a hidden error on the upper levels which is the worst kind of
error. Therefore, always make sure to explicitly convert to
strings (i.e. call the string methods of) all objects used in
a string method.