Add a proxy API for the Pyro4 library so that all GuiBender objects are serializable without pickle (with serializers like serpent and json).
The API uses the Pyro4 autoproxy functionality in order to replace the
actual objects with proxy objects which are perfectly serializable. This
saves us the efford of defining serialization method for every object
we have and preserves the original GuiBender classes untouched. However,
there are still two points to clear out:

1) Object cleanups are no longer automatic as they represent connections
   so we need to perform manual cleanup.
2) Our custom errors are still not passed properly (not serialized) although
   they are still completely readable when raised.