Fix the multi-OS behavior for the temporary file of the desktop dump.
Windows supports only one opening of a file object while UNIX allows
for two. Previous solution was to close the file object as it didn't
seem to be used anywhere throughout the code. As it turns out, it is
used by the destructor in order to destroy the temporary file. The
final solution proposed by Tom is to close it manually and fits best.