Add some default values to the Location class to facilitate its use at least a bit.
This class might be subject to future removal if it does not grow
in complexity. All it does right now is making it really difficult
to manually operate with GuiBender since it requires extra steps of
importing the Location class and initializing when all one wants is
to test something like 'guibender.hover(x,y)'.