Perform large scale key refactoring and adaptation.
There are multiple motivations behind the changes in this commit
which include complete rewriting and separation between the key
pressing (simultaneous, characters only) and text typing (consecutive,
string specialized) variants as well as between the region
(list, integer, string type handling) and the desktop_control
(specilized in backend selection and handling) modules:

- The functionality between the 'press' and 'type' methods is
really mixed between simultaneous and consecutive tapping mode, at
previously focused element or at a match on the screen, of text
input and special key combinations, etc. It then becomes impossible
to have any preference from outside.

- The new backend keymaps involve strings instead of integers (the
autopy case) to represent special keys which challenges us further
to recognize what is meant as a text and what as a special key.

- The naming of the 'press' and 'type' must be similar and since
'type' is a reserved word, it's better to stay safe and avoid it
for something like 'type_text' and 'key_press'.