Add inheritance of backends.
This is a crucial step towards working on multiple parallel screens
which don't necessarily have the same operating system and/or graphical
environment. Both the desktop control and computer vision backends are
no longer recreated on initialization of each region/subregion but
reusable if stored from the end user and inherited by subregions, i.e.
regions created from another region. In this way we can fully operate
on a region with one GUI while doing the same with another using
different CV algorithms and/or screens.