Synchronize the mouse pointer with the one currently on screen.
If this is not done, in some guest OS there is a chance of associating
the last known position of the pointer as (0,0) resulting in any
mouse position being displaced by a fixed distance.

NOTE: Autopy already supports retrieving of the previous pointer state
so it its pointer is already synced.