hybrid: Fall back to template matching in case the feature matching fails
This is about more than feature-poor results. The usual expectation is
that the feature matching is a lot less picky than the template matching
so it doesn't make sense to say there is no match if the template stage
found it but the feature stage refuses it. This further conforms to the
idea of the hybrid method - to be picky at the first stage and less so
at the second.

Template similarity is preferred if it is higher than the feature
similarity since it is more reliable (but more selective), however
also in this case we have to satisfy the final (and not front)
similarity requirement.