Extend all screen targets to a hierarchy of possibilities
This includes a minimal matcher of the first non-image type, namely an image set
where a match is obtained if at least one of the contained images is matched.

NOTE: Initializing an "abstract" class as the base image finder is useless in
all cases and using a fully implemented one instead helps reducing configuration
overhead in terms of number of lines of calling code.