Localize all class-specific dependencies for the CV backends
This makes it possible that a user of only one particular CV backend
should only take care of dependencies for that CV backend. The setup
steps for these dependencies are described in a new document for
different systems like CentOS, Fedora, and Ubuntu.

Previous image preprocessing has been integrated for explicitness
and readabiltity with PIL specific calls where OpenCV can be avoided
as a backend.