Simplify the interface for finding images to returning multiple matches
If a single match is wanted, one could just take the first match.
The single/multiple match separation is kept the same in the Region
interface.

By letting the image matchers return actual match objects, we allow
them to provide with additional useful information such as the size
(width and height) of the detected match and avoid circular importing
which was an old problem with the region module.