Use the OCR backend for all unit tests based on a Qt4 GUI window
The window designed to test various region actions produces a GUI
that looks differently according to platforms and versions so
increasing the robustness to a differently looking GUI control is
an all time must.

Tesseract still shows some weakness at a few spots and reduces the
similarity somewhat but it is still above 0.6 which is much better
than before.