Return sorted matches according to and with included similarity
This could be very useful and can replace the implementation specific
order of each matcher with a standard match order to expect on the other side.

It also helps us make any previous object sampling much more intuitive.
The previous solution for sampling similarity of a target image was more
of a hack and worked only for some methods. The current way to implement
this is more natural and was possible due to the matchers returning more
than just a list of locations (which can also contain the similarity for
each case).