Standardize the synchronization of all CV backends
Due to the peculiar API of OpenCV's feature matchers, we would
typically create the backends during configuration and reuse them
for the actual synchronization. This however, does not conform to
the expectation of how a typical CV backends is configured with
parameters only to be created and then synchronized in order to be
reflected in the backend. The configuration-synchronization passage
is now standardized in all CV backends to do exactly this.