Improve performance a bit by not copying the best found parameters
Since all the modification is already made on the provided parameters
and the return copy is not used used in any other way, it makes better
sense to just calibrate the original parameters instead.

We now only return the maximized similarity from calibration and just
the error during the optimization process (parameters excluded).