Include the RST files generated by sphinx-apidoc
We need these to be available for RTD (Read The Docs) which only
invokes the sphinx-build. We still have to manually regenerate
the RST modules when changing modules but we have a shell script
for this. Placing the RST files into a source directory is required
by RTD in order to avoid a 404 error on the website.