Generalize the border preprocessing to normal, inverted, and no binarization
The border that was previously padded to the text images was fixed to black
thus becoming unusable for black on white background threshold as well as
cases where no binarization is performed. This has the additional advantage
that the border is directly included rather than padded to magnified image
thus saving some performance.