Reduce the default calibration max attempts from 10 to 3
Let us not startle the user that is calibrating for first time -
they wouldn't like losing control for too long when they don't
know how it works and are just probing.

This is also useful for a shorter unit tests that validate the
most meaninful parts of the calibrator. The rest is giant parameteric
spaces for the user to expore.