Scale back the testing surface of the test finder benchmarking to finish early
A full benchmarking with just 3-step calibration and a 1-step search can take
up to 40 minutes so remove some unimplemented or redundant backend combos to
reduce it to 9 seconds. However, this is not a long term solution and if some
of the excluded backends turn out to be more important and changing behavior
too much, we will have to include them back in the future. The same applies
for a general increase in the variety of backends we can use later on.