Replace matched text with fixed locations for all region control tests
Having differently rendered GUI due to OS and Qt version differences is
a problem we generally handle with text matching and OCR. However, this
is error prone and adds complete OCR dependency spectrum while testing
the seemingly irrelevant region control methods like clicking, typing,
etc. So let's fix the locations on the GUI and isolate the testing to
the original target functionality.