Keep one unit test as expected failure due to a Qt race condition
Qt fails to register a close event within a docker container while
it manages to do so outside of it on a system with the same version.
At the same time it may or may not fail to register the close event
(from ALT+F4) on the container-external system as well.