Update the Tesseract OCR backend and data to 4.0.0(.beta1)
In this way move the OCR functionality testing from Fedora 28
to Ubuntu Bionic where the 4.0.0.beta1 is already available
(also through OpenCV).

The current OpenCV package on Fedora either has incompatible
Tesseract version and worse yet - it is missing the entire
OpenCV text module - so we disable OCR testing there at the
time of this commit.