Ultimately disable OpenCV for the PyPy3 variant
Even though most of our backends are based on OpenCV, it is still
not an obligatory dependency (the reason why we can build RPMs
without OpenCV support and don't list it as a main dependency
during packaging in comparison to Python's Pillow).

We generally reuse the LEGACY_OPENCV setting which is no longer
needed and now allow only DISABLE_* switches for any unit tests.