Add TemporaryConfig
Sometimes we need to temporarily change a config item for a few
operations. Repeating this pattern adds extra clutter to the code
and is error prone, since we need to store the old value and
restore it later on.

The TemporaryConfig class extends the original GlobalConfig object
adding context capabilities. In short, once we enter the context
and get an instance of this class, we can change it at will as any
change will be reverted after we exit the context.