Add more clarification regarding QEMU click handing
It was thought that vncdotool was to blame for not correctly
handling click events sent to fast, without a break between them.
However, after investigation this seems to be caused by QEMU's
own rfb implementation. When a custom VNC server installed on the
VM (TightVNC, for example) was used the bug didn't occur.