Support temporarily replacing search paths
The current file resolver uses a shared list of paths, which
might lead to conflicts when searching for files as the caller
has no control over which paths are added to this list.

With a custom file resolver we can have a controlled context
inside which only the added paths will be searched, adding
the ability to use shorter names without the fear of conflicts.