Use the least requiring (sophisticated) text detection by default
In particular, we use contours to analyze for text and the user
could switch to ERStat or EAST if they want to deal with more
elaborate dependencies.

This new requirement needs a slight increase in tolerance for the
default text detection configuration which shouldn't harm any
pre-existing matching cases anywhere.