Cache the large R-CNN models according to architecture or checkpoint
This saves us additional loading and memory usage for each finder
as match settings instance that would otherwise initialize its own
copy of a model of the same architecture or checkpoint.

We can now also reuse models across pattern targets this way using
the "model" CV parameters to point to the same checkpoint file or
the same architecture if no checkpoints are specified.