Allow custom classes for all R-CNN like buttons or GUI elements
Such classes are beyond the COCO dataset and can be trained with
the synthetic dataset in the previously provided example.

As GUI elements and general custom classes may require deeper
training we now allow for full customization of the R-CNN
architecture but keep the current transfer learning configuration
(pretrained and fairly frozen backbone with retrained head).