Extend the documentation-focused README for a more complete usage
The usage example on the web page is too short for some people while
using the API is too long for others. Since the RTD README copied
the project README completely only converting from MarkDown to ReST
and thus having redundancies and frequent update lagging, it is
better to simply repurpose it as middle-ground documentation.