Add support for python 3.9 with disabled AutoPy for it
This includes dependency updates that are minimal for Python 3.9.

This is the final tipping point with AutoPy being stuck in a
distant nightly build of Rust and not compiling for Python 3.9
for more than a year now. Our next step will be to move to the
more popular and actively maintained PyAutoGUI.