Adapt the pip deps on Windows to our standard requirements file
Also disable and update the versions of some otherwise unbuildable
pip deps on Windows.

In particular, refactor together all dependencies for just testing
and fix the versions to the closes working ones per platform. E.g.
newer PyQt5 is broken on Linux but older one is broken on Windows
in between the two selected versions.

The pip upgrade is restricted to the local users because Windows
emits error messages (nonfatal but still better to only allow
warnings). This differs a bit from the other packager scripts but
we can later on adapt them to user-local installation too if this
ends up being more preferred.

Optional proxy interface have been generalized and OCR on Windows
is not yet available due to no official Tesseract OCT installers.