Rewrite a unit test that should focus on the steps file's interface
Often when writing unit tests we have to think a lot not just about
what to assert but also what not to assert. Not asserting means
abstracting away and allowing variation. Too many assertions and the
tests become fragile about irrelevant changes while they should in
fact be as fragile as possible about the validated functionality
they should intentionall fail for.