Allow text targets to be formally constructed from a text file
The previous approach towards using a text file with a different
name than the initial 30 characters (or less) of the text target
was too implicit and involved secondary access to a public attribute,
something we used to do to preserve greater API uniformity and
always assume file-less text targets unless otherwise handled.

In the new approach we will provide a file if we mean to use a
custom file with a different name to load the text value (data)
from and provide a text value if we mean file-less or auto-detected
text file usage.

We also explicitly protect text value which is is a filename to
end up in looking up text files in subdirectories, this can have
unforeseen effects and must not be supported.