Provide the key and mouse button maps for the simple interface
These previously required an manual extra import of the inputmap
module, something that is hardly "simple" enough to quality for
ways of using this interface or a region to be obtained first in
order to reuse its keys (also hardly "simple"). With the current
change we will provide somewhat similar behavior to the proxy and
main guibot interfaces and allow the user to be provided with all
basic necessities when using just the simple interface.

As we cannot simply assign to an already imported variable within
the namespace of the simple API we make use of a named tuple to
provide the next easiest option for accessing all input maps.