Add a setting to wait for animations and thus moving targets
This is useful when clicking, typing, or many other actions that
need the found match and location to click on to coincide, i.e.
the target should say where it is on the screen and not move.

Integration testing here requires a advances in our overall testing
environment as we need to simulate moving objects that stop and
are then clicked so instead we move further towards isolation testing
mocking and providing changing match coordinates - fewer dependencies
and requirements on complex integration testing environments.