Return region (self) also in the case of wait_vanish
The behavior of wait_vanish was inconsistent with all other region
methods which return either a match (a subregion with extra match
data) or themselves if no matching is done, all to allow a chain
of calls with incrementally smaller region scopes.

The same old behavior is also too complex in the sense that it
would either return a boolean if it succeeds or raise an error
thus signalling success in two separate ways. The function should
instead eitehr throw an error to signal failure or ignore errors
altogether and return a boolean status instead. Let's choose the
first which is more consistent with other methods again.