Add expectation-enhanced methods for key pressing
It is also possible to press a key and expect a target to appear
or disappear as GUI behaviors and not just click a mouse button.

We also properly restrict the treatment of the expected target
and its allowed types as not all of the previously specified ones
would have worked before.