Add ability to retry an action when expected event does not happen
This is similar to human behavior with GUIs when a human clicks a
mouse button or presses a key and expects an event but not seeing
the event after a while will often make the human repeat their
previous action, i.e. click or press again.

This is especially useful if any of the display controller backends
misbehaves, e.g. drops a click or a key press due to a race condition,
internal bugs or others. Therefore, the automation task here is one
where a program is adapted to behave in a rational way with respect
to a misbehaving interface and not the other way around as one may
not have any control over the interface and/or display controller
communication with the OS. While many use cases of guibot are to
test and find problems with a given GUI application, there are still
cases where a test might fail due to bugs or problems in dependencies
that are outside of the GUI application developer's scope and would
benefit from such retries.