Add call delegation tests for both guibot extension interfaces
This requires a bit of meta-programming and other mock flexing
but it provides complete coverage of API consistency and also:

1) Already detected some stale out of date calls for two functions.

2) Already suggested improvement on the design of the interfaces
by using generic positional and keyword arguments.

The generalization of arguments to pass along will also simplify
adding new interface methods in the future as by the end of this
branch we added multiple such and accumulate the need to simplify.