Drop requirement on using tessdata environment variable
As we now require additional tessdata than just the trained languages
and guibot's misc data path is not included in all forms of packaging
it is better to drop any tessdata distributed within guibot and thus
any requirement on explicitly providing a (tess)data path in the form
of CV parameter or an environment variable. These would still take
increasing order of precedence if set though to support any previous
customization.