Update RTD documentation with recent developments
RTD configuration file and a default index.rst is now required and
some Sphinx settings have also changed accordingly. Let us also use
the opportunity to simplify the API doc generation bash script.

The README backtick links had to be toned down a bit in order to
have a working conversion into reStructuredText and the docs build
locally without any warnings but RTD seems to misbehave detecting
the index.rst file and complaining it is not in any toctree.