Allow for retries also within the select-at convenience method
The additional resilience of the `*_expect` methods can also be
used in the particular case of select-at using an image within a
dropdown menu in the case where the click might be unsuccessful
and the dropdown menu will not open as a result.

Use the opportunity to reset cursor location (important if it is
visible on the screen) in all these retry-capable cases.