Fix runaway dependencies that our fixed dependencies don't control
In particular lock the numpy version to avoid an update to a major
2.0 yet and incremental version to previously known working one.

Even though we don't explicitly control the numpy version (but
implicitly via other dependencies that we hope have overall sane
requirements) a recent release of NumPy 2.0 breaks the way it is
used in OpenCV and PyTorch which at the same time don't demand a
safe version (and rather seem to download the latest one). The
way around this is to do some of their work for them and require
a sane version that their fixed and controlled version will still
work with. Same applices for vncdotool's Twisted dependency.