Add CI for auto release and push artifacts with it (.rpm and .deb)
This commit adds a new CI job to push artifacts and generate a new
release on every update we make. It will make the process after
running unit tests so that we can garantee a safe artifact for the
final user with our latest changes. It was needed to change a few
small things on the packagers (deb and rpm) so that we keep the
artifacts afer installing them on the docker containers.