Add push artifacts functionality to our CI (.rpm and .deb)
This commit adds a new CI job to push artifacts to GitHub.
After generating them with the deb and rpm packagers and fully
testing.

A VERSION file was added to track version numbers.
Small changes to the packagers themselfs have been made to
not delete the artifacts that they generate.