Add a test skip for tests of feature to avoid OOM error
This commit adds a simple skip for some test variants, so that we
avoid OOM errors on machine with less than 16gb of RAM. The error
seems to happen whenever we use BRISK and BruteForce* in some
specific scenarios. Test feature nomatch seems to work with some
combinations however, test feature same is much less permissive
with errors for every combination of BRISK and Brute-Force.